fix(#3060): split model_loader_dialog.py — separate dialog, loader, and tree widget
- Extracted model load/download actions and info formatters into ModelLoaderMixin (model_loader.py)
- Extracted repo/embedded/community/imported tab builders into ModelTreeMixin (model_tree_widget.py)
- ModelLoaderDialog slimmed to thin orchestrator using cooperative multiple inheritance
- Backward-compat preserved: ModelLoaderDialog still importable from model_loader_dialog.py
